Extended rate limits to add in conditional rate limits
