+++ kytos/core/buffers/factory/__init__.py
@@ -1,0 +1,5 @@
+"""Factory for KytosEventBuffer"""
+
+from .buffers import buffer_from_config
+
+__all__ = ('buffer_from_config',)
